+++ example.js
@@ -4,4 +4,20 @@ const pwds = require('./index');
 pwds.generate().forEach(pw => console.log('default', pw));
 
 // generate passwords with custom parameters
-pwds.generate({ quantity: 21, pwLength: 16 }).forEach(pw => console.log('custom', pw));
+const params = {
+  quantity: 20,
+  pwLength: 32,
+  uppercases: 'abcdef0123456789',
+  lowercases: 'abcdef0123456789',
+  symbols: 'abcdef0123456789',
+  digits: 'abcdef0123456789',
+};
+pwds.generate(params).forEach(pw => console.log('custom', pw));
+
+// generate one password with default parameters
+const password = pwds.password();
+console.log('one pwd', password);
+
+// generate passwords with custom parameters
+const customParPassword = pwds.password(params);
+console.log('one custom pwd', customParPassword);