+++ src/pwdgen.js
@@ -1,3 +1,17 @@
+/**
+ * @fileoverview password generator engine
+ *
+ * @description
+ * This file contains the password creation engine. The internal
+ * state consists of the characters used to construct passwords
+ * and related restrictions. The module exports two functions:
+ * 1) `generate`: returns an array of strings/passwords;
+ * 2) `password`: returns a single string/password
+ *
+ * @author Thorn Duke
+ * @copyright 2024
+ * @license GPL-3.0-or-later
+ */
 'use strict';
 
 const crypto = require('node:crypto');
@@ -5,17 +19,30 @@ const crypto = require('node:crypto');
 const $_DEBUG = false;
 
 let defaults = {
-  pwLength: { type: 'number', min: 8, max: 128, default: 12 },
-  quantity: { type: 'number', min: 1, max: 100, default: 10 },
-  uppercases: { type: 'string', min: 4, max: 128, default: 'ABCDEFGHIJKLMNOPQRSTUVWXYZ' },
-  uppercasesQty: { type: 'number', min: 1, max: 2, default: 1 },
-  lowercases: { type: 'string', min: 4, max: 128, default: 'abcdefghijklmnopqrstuvwxyz' },
-  lowercasesQty: { type: 'number', min: 1, max: 2, default: 1 },
-  digits: { type: 'string', min: 4, max: 128, default: '1234567890' },
-  digitsQty: { type: 'number', min: 1, max: 2, default: 1 },
-  symbols: { type: 'string', min: 4, max: 128, default: '£$%&+*/-@#' },
-  symbolsQty: { type: 'number', min: 1, max: 2, default: 1 },
+  pwLength: { type: 'number', min: 8, max: 128, value: 12 },
+  quantity: { type: 'number', min: 1, max: 100, value: 10 },
+  uppercases: { type: 'string', min: 4, max: 128, value: 'ABCDEFGHIJKLMNOPQRSTUVWXYZ' },
+  uppercasesQty: { type: 'number', min: 1, max: 2, value: 1 },
+  lowercases: { type: 'string', min: 4, max: 128, value: 'abcdefghijklmnopqrstuvwxyz' },
+  lowercasesQty: { type: 'number', min: 1, max: 2, value: 1 },
+  digits: { type: 'string', min: 4, max: 128, value: '1234567890' },
+  digitsQty: { type: 'number', min: 1, max: 2, value: 1 },
+  symbols: { type: 'string', min: 4, max: 128, value: '£$%&+*/-@#' },
+  symbolsQty: { type: 'number', min: 1, max: 2, value: 1 },
 };
+
+function resetDefaults() {
+  defaults.pwLength.value = 12;
+  defaults.quantity.value = 10;
+  defaults.uppercases.value = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ';
+  defaults.uppercasesQty.value = 1;
+  defaults.lowercases.value = 'abcdefghijklmnopqrstuvwxyz';
+  defaults.lowercasesQty.value = 1;
+  defaults.digits.value = '1234567890';
+  defaults.digitsQty.value = 1;
+  defaults.symbols.value = '£$%&+*/-@#';
+  defaults.symbolsQty.value = 1;
+}
 
 /**
  * Produces a random integer between `minInteger` (inclusive)
@@ -41,7 +68,7 @@ function getRandomInt(minInteger, maxInteger) {
  * Uses the [Fisher-Yates Shuffle Algorithm](https://en.wikipedia.org/wiki/Fisher%E2%80%93Yates_shuffle)
  * applied three times.
  *
- * @param {Array} arr An array of elements
+ * @param {Array} arr - An array of elements
  * @returns {Array} A shuffled array of elements
  */
 function shuffleArray(arr) {
@@ -88,10 +115,10 @@ function shuffleString(str) {
  * @example createPool() // '5%k#BC1j2Fm3GEiADh6l@&0'
  */
 function createPool() {
-  const uc = shuffleString(defaults.uppercases.default);
-  const lc = shuffleString(defaults.lowercases.default);
-  const d = shuffleString(defaults.digits.default);
-  const s = shuffleString(defaults.symbols.default);
+  const uc = shuffleString(defaults.uppercases.value);
+  const lc = shuffleString(defaults.lowercases.value);
+  const d = shuffleString(defaults.digits.value);
+  const s = shuffleString(defaults.symbols.value);
 
   const result = shuffleString(uc + lc + d + s);
 
@@ -115,7 +142,7 @@ function createPool() {
  * @param {string} refStr the string used to check the password
  * @returns {boolean}
  *
- * @example checkPassword('fT5s@hjJ', 2, defaults.uppercases.default) // true
+ * @example checkPassword('fT5s@hjJ', 2, defaults.uppercases.value) // true
  */
 function checkPassword(str, occurrences, refStr) {
   let result = false;
@@ -151,10 +178,10 @@ function isValidPassword(pwd) {
   let result = true;
 
   if (
-    !checkPassword(pwd, defaults.uppercasesQty.default, defaults.uppercases.default) ||
-    !checkPassword(pwd, defaults.lowercasesQty.default, defaults.lowercases.default) ||
-    !checkPassword(pwd, defaults.digitsQty.default, defaults.digits.default) ||
-    !checkPassword(pwd, defaults.symbolsQty.default, defaults.symbols.default)
+    !checkPassword(pwd, defaults.uppercasesQty.value, defaults.uppercases.value) ||
+    !checkPassword(pwd, defaults.lowercasesQty.value, defaults.lowercases.value) ||
+    !checkPassword(pwd, defaults.digitsQty.value, defaults.digits.value) ||
+    !checkPassword(pwd, defaults.symbolsQty.value, defaults.symbols.value)
   )
     result = false;
 
@@ -174,7 +201,7 @@ function createPassword() {
 
   do {
     result = '';
-    for (let i = 1; i <= defaults.pwLength.default; i++) {
+    for (let i = 1; i <= defaults.pwLength.value; i++) {
       const index = getRandomInt(0, pool.length - 1);
       result += pool[index];
     }
@@ -189,15 +216,13 @@ function createPassword() {
 
 /**
  * Produces an array containing passwords of length equal
- * to _length_ characters. the array has `getPwdListLength`
- * elements. It is the method called by the extension
- * to produce and print password lists.
+ * to _defaults.pwLength.value_ characters.
  *
  * @returns {string[]} An array of strings
  */
 function createPasswordList() {
   let result = [];
-  const spins = defaults.quantity.default;
+  const spins = defaults.quantity.value;
 
   for (let i = 1; i <= spins; i++) {
     const password = createPassword();
@@ -218,95 +243,39 @@ function createPasswordList() {
  * @param {object} [params] An object containing parameters
  */
 function checkParams(params) {
-  function isConform(val, reqType, min, max) {
+  function isConform(val, { type, min, max }) {
     return (
       val !== undefined &&
-      typeof val === reqType &&
-      (reqType === 'number' ? Number.isInteger(val) === true : true) &&
-      (reqType === 'number' ? val >= min && val <= max : val.length >= min && val.length <= max)
+      typeof val === type &&
+      (type === 'number' ? Number.isInteger(val) === true : true) &&
+      (type === 'number' ? val >= min && val <= max : val.length >= min && val.length <= max)
     );
   }
 
-  if (
-    isConform(
-      params.pwQuantity,
-      defaults.quantity.type,
-      defaults.quantity.min,
-      defaults.quantity.max
-    )
-  )
-    defaults.quantity.default = params.pwQuantity;
+  if (isConform(params.quantity, defaults.quantity)) defaults.quantity.value = params.quantity;
 
-  if (
-    isConform(params.pwLength, defaults.pwLength.type, defaults.pwLength.min, defaults.pwLength.max)
-  )
-    defaults.pwLength.default = params.pwLength;
+  if (isConform(params.pwLength, defaults.pwLength)) defaults.pwLength.value = params.pwLength;
 
-  if (
-    isConform(
-      params.uppercases,
-      defaults.uppercases.type,
-      defaults.uppercases.min,
-      defaults.uppercases.max
-    )
-  )
-    defaults.uppercases.default = params.uppercases;
+  if (isConform(params.uppercases, defaults.uppercases))
+    defaults.uppercases.value = params.uppercases;
 
-  if (
-    isConform(
-      params.uppercasesQty,
-      defaults.uppercasesQty.type,
-      defaults.uppercasesQty.min,
-      defaults.uppercasesQty.max
-    )
-  )
-    defaults.uppercasesQty.default = params.uppercasesQty;
+  if (isConform(params.uppercasesQty, defaults.uppercasesQty))
+    defaults.uppercasesQty.value = params.uppercasesQty;
 
-  if (
-    isConform(
-      params.lowercases,
-      defaults.lowercases.type,
-      defaults.lowercases.min,
-      defaults.lowercases.max
-    )
-  )
-    defaults.lowercases.default = params.lowercases;
+  if (isConform(params.lowercases, defaults.lowercases))
+    defaults.lowercases.value = params.lowercases;
 
-  if (
-    isConform(
-      params.lowercasesQty,
-      defaults.lowercasesQty.type,
-      defaults.lowercasesQty.min,
-      defaults.lowercasesQty.max
-    )
-  )
-    defaults.lowercasesQty.default = params.lowercasesQty;
+  if (isConform(params.lowercasesQty, defaults.lowercasesQty))
+    defaults.lowercasesQty.value = params.lowercasesQty;
 
-  if (isConform(params.digits, defaults.digits.type, defaults.digits.min, defaults.digits.max))
-    defaults.digits.default = params.digits;
+  if (isConform(params.digits, defaults.digits)) defaults.digits.value = params.digits;
 
-  if (
-    isConform(
-      params.digitsQty,
-      defaults.digitsQty.reqType,
-      defaults.digitsQty.min,
-      defaults.digitsQty.max
-    )
-  )
-    defaults.digitsQty.default = params.digitsQty;
+  if (isConform(params.digitsQty, defaults.digitsQty)) defaults.digitsQty.value = params.digitsQty;
 
-  if (isConform(params.symbols, defaults.symbols.type, defaults.symbols.min, defaults.symbols.max))
-    defaults.symbols.default = params.symbols;
+  if (isConform(params.symbols, defaults.symbols)) defaults.symbols.value = params.symbols;
 
-  if (
-    isConform(
-      params.symbolsQty,
-      defaults.symbolsQty.type,
-      defaults.symbolsQty.min,
-      defaults.symbolsQty.max
-    )
-  )
-    defaults.symbolsQty.default = params.symbolsQty;
+  if (isConform(params.symbolsQty, defaults.symbolsQty))
+    defaults.symbolsQty.value = params.symbolsQty;
 
   if ($_DEBUG)
     console.log('| DEBUG | <checkParams> =>', {
@@ -318,5 +287,13 @@ function checkParams(params) {
 module.exports.generate = function (params) {
   if (params) checkParams(params);
   const passwordList = createPasswordList();
+  resetDefaults();
   return passwordList;
+};
+
+module.exports.password = function (params) {
+  if (params) checkParams(params);
+  const pwd = createPassword();
+  resetDefaults();
+  return pwd;
 };